Refactor YouTube service utilities into a separate module; update Dockerfile to optimize package installation.
